Example of Core Animation API
This is a Xcode project which contains examples of Core Animation API
or the Quartzcore framework in iOS.
These are written in swift programming language.
Examples here includes the explanation of animations like rotation of
sublayer on the container layer, make an object move on a specific path
(created by UIBezierPath), animating more than one animation
simultaneously using CAAnimationGroup and a small example of password
object vibrating when it has the wrong value.

This was actually made for the demonstration for Grad Presentation on
the topic “Core Animation”